fix: 코드 분리 및 수정
