+++ src/lib/utils/public/screenDimensions.ts
@@ -1,0 +1,3 @@
+import { Dimensions } from 'react-native'
+
+export const { width: SCREEN_WIDTH, height: SCREEN_HEIGHT } = Dimensions.get('window');
